requirements.txt
Command==0.1.0
python-dotenv==0.21.0
